requirements.txt
anyio==4.9.0
ariadne==0.26.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.1
colorama==0.4.6
deepmerge==2.0
Flask==3.1.1
flask-cors==6.0.1
flask-redis==0.4.0
Flask-SocketIO==5.5.1
flask-talisman==1.1.0
graphql-core==3.2.5
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
selenium==4.21.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
